Refactor login method from user class
The login method from user class used to be including the logic of navigating to user menu, which is inappropriate compared to placing it in the MainForm class. There's also new helper class added to help with the login method or upcoming method in this project that will return many different types of value.